% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft helpers.R
\name{compute_dispersion_percentage}
\alias{compute_dispersion_percentage}
\title{Simulate number of secondary infections for given dispersion}
\usage{
compute_dispersion_percentage(disp, R = 1, sample.size = 1000, prob = 0.2)
}
\arguments{
\item{disp}{dispersion}

\item{R}{reproductive number}

\item{sample.size}{sample size for simulation}

\item{prob}{ratio of primary cases under consideration. 20% is standard.}

\item{chain}{length of chain

Compute percentage of secondary infections from upper percentage of primary cases}
}
\description{
Simulate number of secondary infections for given dispersion
}
