% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags helpers.R
\name{extract_mcmc}
\alias{extract_mcmc}
\title{Extract MCMC output}
\usage{
extract_mcmc(
  name,
  dim_names = NULL,
  quantiles = F,
  samples = F,
  df = NULL,
  add_info = TRUE,
  t_only_model = FALSE,
  sample_name = NULL,
  sample_age = NULL,
  without_name = NULL,
  exact_name = FALSE
)
}
\arguments{
\item{name}{name of parameter}

\item{dim_names}{vector of dimension names that were extracted}

\item{quantiles}{extract mean and sd or quantiles (default FALSE, which is mean and sd)}

\item{df}{data.frame optional ,default is res in parent environment}

\item{add_info}{adds automatically date to t from data and country/region name from colnames of d}

\item{sample_name}{subsample of name observations to look at. logical takes 16 random. vector of characters takes those. number takes that number at random.}

\item{without_name}{substrings that should not be taken from mcmc output}

\item{exact_name}{if name of parameter has to match exactly. standard FALSE.}
}
\value{
data.frame
}
\description{
Extract MCMC output
}
