% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags helpers.R
\name{pull_mcmc_sample}
\alias{pull_mcmc_sample}
\title{pull sample of one variable of mcmc draws}
\usage{
pull_mcmc_sample(var_name)
}
\arguments{
\item{var_name}{name of variable}
}
\value{
vector of draws
}
\description{
pull sample of one variable of mcmc draws
}
