% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags helpers.R
\name{show_effects_sample}
\alias{show_effects_sample}
\title{Plot covariate effects according to mcmc output}
\usage{
show_effects_sample(
  cov.choice = macro$cov,
  times = FALSE,
  filter_age = NULL,
  average.age = FALSE,
  hlines = TRUE,
  CI_large = TRUE,
  add_prior = FALSE,
  background_prior = TRUE,
  scale = "free_y",
  order = FALSE,
  width_errorbar = 1,
  ...
)
}
\arguments{
\item{times}{if numeric, multiplicates output by times, if function applies function}

\item{...}{}

\item{dim_names}{dimension names of effects (standard is age and m, which is the number of covariates)}
}
\description{
New version of plot_effects
}
