% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags helpers.R
\name{show_total_effect}
\alias{show_total_effect}
\title{Plot combined effect of weather on time}
\usage{
show_total_effect(
  covariates = macro$cov.weather,
  forecast = FALSE,
  CI_large = TRUE,
  hlines = TRUE,
  draws_max = NULL,
  choose_age = FALSE,
  choose_c = NULL,
  average_c = FALSE,
  average_age = FALSE,
  smooth = FALSE,
  show_covariate = FALSE,
  ...
)
}
\arguments{
\item{forecast}{standard FALSE, if TRUE takes weather variables for whole year and averages for each day in year}

\item{hlines}{adds hline at zero}

\item{draws_max}{maximum number of draws from posterior for speed, standard 100}

\item{choose_age}{subset of ages to look at}

\item{choose_c}{subset of units (countries) to look at}

\item{average_c}{show all results or only average for all countries}

\item{...}{further parameters for extract_mcmc}

\item{grid}{return grid with name on age instead of standard which is one single plot for each name}

\item{model}{standard FALSE, if TRUE uses linear model, if FALSE uses kernel sieve}

\item{wrap}{if TRUE uses wrap instead of grid plotting ages in one plot}

\item{dim_names}{dimension names for effects standard age with m for covariate}
}
\value{
ggplot
}
\description{
Plot combined effect of weather on time
}
