% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft helpers.R
\name{total_diff_effect}
\alias{total_diff_effect}
\title{Compute difference in reproduction rate between two inputs}
\usage{
total_diff_effect(
  df.values,
  choose_age = NULL,
  average = FALSE,
  cov.cur = macro$cov
)
}
\arguments{
\item{df.values}{data.frame with name and value1 and value2 of covariate}

\item{choose_age}{if age should be filtered}

\item{average}{if output should be weighted average across age groups}

\item{cov.cur}{covariates that one can choose from (standard only continuous)}
}
\description{
Compute difference in reproduction rate between two inputs
}
